% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNSMILER_Public_InfVis.R
\name{calcxbar}
\alias{calcxbar}
\title{calcxbar}
\usage{
calcxbar(posterior, dbn, pctiles)
}
\arguments{
\item{posterior}{- data frame formatted as per dbn$posterior (see DBNSMILER-package 
documentation); $node,$state,$t,$p,$scenario,$pctile,$xbar}

\item{dbn}{DBNSMILER object (see DBNSMILER-package documentation). Use dbn to match node 
names to node state thresholds & states needed to determine expected value assuming uniform
distributions.}

\item{pctiles-}{numeric vector of percentiles for evaluating expected value=weighted mean at}
}
\value{
posterior data frame (see DBNSMILER-package documentation) with expected val/weighted
mean results added; if error, original posterior data frame is returned
}
\description{
Calculates the expected value from posterior probabilities assuming a uniform
distribution across states, at specified percentiles. The results are added to the 
posterior object.
}
\examples{
none
}

