% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNSMILER_Public_Get_Set.R
\name{chardate2num}
\alias{chardate2num}
\title{chardate2num}
\usage{
chardate2num(strdate, timeunit)
}
\arguments{
\item{strdate}{is a CHARACTER string of a POSIXlt date; can handle vector of strdates}

\item{timeunit}{- 'year','month','day','hour',minute','second'}
}
\value{
numeric date - number of timeunits since 1900/1/1
}
\description{
This function converts a R char date string into numeric time in terms of 
#timeunits since 1900/1/1, where timeunits are: years, months, days, hours, minutes, seconds
}
\examples{
none
}

