% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNSMILER_Public_InfVis.R
\name{clearevid}
\alias{clearevid}
\title{clearevid}
\usage{
clearevid(dbn)
}
\arguments{
\item{dbn}{- a DBNSMILER object}
}
\value{
1 if success, -1 if fail
}
\description{
Clear all evidence in the SMILE object
DO NOT USE - SMILE THROWS EXCEPTIONS FOR CLEARTEMPORALEVIDENCE
}
\examples{
none
}

