% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNSMILER_Public_InfVis.R
\name{dbnplot}
\alias{dbnplot}
\title{dbnplot}
\usage{
dbnplot(sellabel, posterior, plottype, dateformat, customiseplots)
}
\arguments{
\item{sellabel}{- either 1) name of node to plot, or 2) ddf key (node+scn)}

\item{posterior}{- data frame formatted as per dbn$posterior (see DBNSMILER-package 
documentation) WITH/WITHOUT $node and $scenario; -$node,$state,$t,$p,-$scenario,$pctile,$xbar}

\item{plottype}{- 'statebeliefs' (state probability plots) OR 'wmean' (weighted mean) OR 'both'}

\item{dateformat}{- char specifying x-axis date format, e.g. \%m-\%Y for months-year
\%d for day, \%H for hour, \%M for minute, \%S for second. Currently only handles dates.}

\item{customiseplots}{- if 1, looks for function customiseplotswrapper to customise plot colours}
}
\value{
a ggplot object
}
\description{
Plot weighted mean or state beliefs or both for specified DBN nodes. Handles both 
the case where posterior is missing $node and $scenario (as part of batch plotting - see 
visbatchdbn) or the normal posterior with $node and $scenario
}
\examples{
none
}

