% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNSMILER_Public_InfVis.R
\name{editevid}
\alias{editevid}
\title{editevid}
\usage{
editevid(action, evid, evidrow, sr)
}
\arguments{
\item{action}{- 'add', 'modify', 'remove', 'mvup', 'mvdown'}

\item{evid}{table (data frame)}

\item{evidrow}{- a vector of char corresponding to a row of the evid table to add/modify}

\item{sr}{- integer of the selected row id}
}
\value{
evid table formatted/edited; returns input arg 'evid' if error check fails
}
\description{
Perform operations on the evidence table:add a row, modifyrow, remove, mvup, mvdown.
Format of evid table: $EvidID (int), $Scenario (char), $Node (name as str), $DateStart (char), 
$DateEnd (char), $SelectState (char), SoftEvidence (char); see DBNSMILER-package documentation
}
\examples{
none
}

