% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNSMILER_Public_Network.R
\name{editrules}
\alias{editrules}
\title{editrules}
\usage{
editrules(action, rules, rulesrow, sr)
}
\arguments{
\item{action}{- 'add', 'modify', 'remove', 'mvup', 'mvdown'}

\item{rules}{table (data frame)}

\item{rulesrow}{- a vector of char corresponding to a row of the rules table to add/modify}

\item{sr}{- integer of the selected row id}
}
\value{
node rules table formatted/edited; returns input arg 'rules' if error check fails
}
\description{
Perform operations on the rules table: add a row, modify row, remove, mvup, mvdown.
Note that the first two rows of rules are header rows. Format of rules table is as described in 
'initrules'
}
\examples{
none
}

