% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNSMILER_Public_Network.R
\name{findnodebyname}
\alias{findnodebyname}
\title{findnodebyname}
\usage{
findnodebyname(dbn, nodename)
}
\arguments{
\item{dbn}{- DBNSMILER structure object}

\item{nodename}{- string array containing the SMILE node name}
}
\value{
index to dbn, -1 if not found
}
\description{
Finds the index to dbn for the specified node
}
\examples{
none
}

