% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNSMILER_Public_Network.R
\name{formatcpt}
\alias{formatcpt}
\title{formatcpt}
\usage{
formatcpt(parnames, parstates, cpt, states)
}
\arguments{
\item{parnames}{- vector of parent node names (char)}

\item{parstates}{- list of vectors of parent state names (char)}

\item{cpt}{- node cpt as numeric matrix}

\item{states}{- vector of node state labels (char)}
}
\value{
node cpt formatted with labels for display purposes
}
\description{
Adds labels to the raw cpt matrix (obtained with getcpt)
}
\examples{
none
}

