% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNSMILER_Public_InfVis.R
\name{gendatetimeevid}
\alias{gendatetimeevid}
\title{gendatetimeevid}
\usage{
gendatetimeevid(evidence, node, chartime, scenario, states, datetype)
}
\arguments{
\item{evidence}{- dbn$evid structure (see def in DBNSMILER-package)}

\item{node}{- char name (id) of datetime node (e.g. 'Time_of_Year')}

\item{chartime}{- vector of chars of datetime label corresponding to time range for simulation
vector length SHOULD be same length as numtslices (see setnumtslices)}

\item{scenario}{- the scenario ID (char)}

\item{states}{(vector of char state labels for the datetime node)}

\item{datetype}{('month' only one currently supported)}
}
\value{
evid with datetime node evidence added; returns evid if function fails
}
\description{
Generate evidence for setting a datetime node in the DBN (e.g. for a 
'Time_of_Year' node and dbn$datetype=months, generate evid=Jan, Feb, Mar etc. foreach month)
}
\examples{
none
}

