% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNSMILER_Public_Get_Set.R
\name{getcpt}
\alias{getcpt}
\title{getcpt}
\usage{
getcpt(net, n, dtslice)
}
\arguments{
\item{net}{- SMILE net object}

\item{n}{- SMILE node id (integer)}

\item{dtslice}{- node dt slice (0 = same time slice, 1 = t+1 etc.)}
}
\value{
node cpt as a numeric matrix
}
\description{
Gets the node cpt table (no labels)
}
\examples{
none
}

