% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNSMILER_Public_Get_Set.R
\name{getmaxdtslices}
\alias{getmaxdtslices}
\title{getmaxdtslices}
\usage{
getmaxdtslices(net)
}
\arguments{
\item{net}{- SMILE net object}
}
\value{
node dt slices as int vector
}
\description{
Gets the max #dt slices in DBN (0=same time slice, 1 = t+1 etc.)
}
\examples{
none
}

