% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNSMILER_Public_Get_Set.R
\name{getnodedtslices}
\alias{getnodedtslices}
\title{getnodedtslices}
\usage{
getnodedtslices(net, n)
}
\arguments{
\item{net}{- SMILE net object}

\item{n}{- SMILE node id (integer)}
}
\value{
node dt slices as int vector (sorted in ascending order)
}
\description{
Gets the node dt slices (0=same time slice, 1 = t+1 etc.)
}
\examples{
none
}

