% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNSMILER_Public_Get_Set.R
\name{getnodename}
\alias{getnodename}
\title{getnodename}
\usage{
getnodename(net, n)
}
\arguments{
\item{net}{- SMILE net object}

\item{n}{- SMILE node id (integer)}
}
\value{
node name as char string
}
\description{
Gets the node name
}
\examples{
none
}

