% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNSMILER_Public_Get_Set.R
\name{getnodestates}
\alias{getnodestates}
\title{getnodestates}
\usage{
getnodestates(net, n)
}
\arguments{
\item{net}{- SMILE net object}

\item{n}{- SMILE node id (integer or vector of integers)}
}
\value{
node states as a vector of chars (if n is one node), or a list of vector of chars (when
n is a vector of nodes) - each list item corresponds to a node in n
}
\description{
Gets the node states - this function can handle vector of nodes or empty n
}
\examples{
none
}

