% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNSMILER_Public_Get_Set.R
\name{getnodetemporaltype}
\alias{getnodetemporaltype}
\title{getnodetemporaltype}
\usage{
getnodetemporaltype(net, n)
}
\arguments{
\item{net}{- SMILE net object}

\item{n}{- SMILE node id (integer)}
}
\value{
integer, 0 = static, 3 = dynamic plate node
}
\description{
Get the temporal type of a node. NOTE: ERROR IS THROWN IF NODE IS NOT-A-PLATE-NODE
}
\examples{
none
}

