% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNSMILER_Public_InfVis.R
\name{getpmass}
\alias{getpmass}
\title{getpmass}
\usage{
getpmass(dbn, P, ind, tind, labelthresh, plotrange)
}
\arguments{
\item{dbn}{is DBNSMILER object}

\item{P}{is posterior data frame (data frame: node=char,state=char,t=char,p=numeric,scenario=int,
pctile=numeric,xbar=numeric expected value for given percentile,ybar=observed numeric exp. val 
for given percentile, py = numeric observed probability) - see DBNSMILER.
ASSUMES that P is sorted in increasing time order
ASSUMES that order of parents in dbn$node$parnodenames = order of parents in $formattedcpt}

\item{ind}{is index to dbn$node[[ind]] for selected node}

\item{tind}{is index to node time slice - e.g. dbn$node[[ind]]rules[[tind]]}

\item{labelthresh}{- (0,1) - max pr threshold for adding annotation with rule id on plot}

\item{plotrange}{- 2-element date/time vector (same datetype as dbn) of form [startdate,enddate]
defining thresholds for plotting}
}
\value{
obj$pmass is data frame with $t = datetime (same units as dbn object), $rule = rule id
$p = probability mass p, $node = nodename; obj$plotobj = ggplot plot object of pmass over time
}
\description{
Get rule 'activation strength'/probability mass over time and plot
}
\examples{
none
}

