% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNSMILER_Public_Get_Set.R
\name{getposterior}
\alias{getposterior}
\title{getposterior}
\usage{
getposterior(net, nodes, nodestates, chartime, scenario, posterior)
}
\arguments{
\item{net}{- SMILE DBN object}

\item{nodes}{- SMILE node id name (char or vector of chars)}

\item{nodestates}{- vector of chars of state labels OR list of vector of chars of state labels}

\item{chartime}{- vector of chars of datetime label - ASSUMES same chartime forall nodes; vector
length MUST be same length as numtslices (see setnumtslices)}

\item{scenario}{- the scenario ID (int) ASSUMES all nodes are for same scenario}

\item{posterior}{- data frame formatted as per dbn$posterior (see DBNSMILER-package 
documentation); $node,$state,$t,$p,$scenario,$pctile,$xbar}
}
\value{
posterior data frame (see DBNSMILER-package documentation); if error, original
posterior data frame is returned, otherwise posteriors are ADDED to original posterior df
}
\description{
Get posterior probabilities for specified nodes and adds results to the end
of a posterior data frame (NOTE: does NOT check for duplicates).
}
\examples{
none
}

