% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNSMILER_Private.R
\name{initdbn}
\alias{initdbn}
\title{initdbn PRIVATE}
\usage{
initdbn(net, n, xdslfilepath)
}
\arguments{
\item{net}{is the SMILE DBN java object}

\item{n}{is the SMILE node ID (integer) to the first node the DBN}

\item{xdslfilepath}{is a char of the filename and path to the SMILE .xdsl file}
}
\value{
void
}
\description{
This function constructs the dbn structure used to store the DBN
}
\examples{
none
}

