% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNSMILER_Public_Network.R
\name{maplingprtopr}
\alias{maplingprtopr}
\title{maplingprtopr}
\usage{
maplingprtopr(linguisticprs)
}
\arguments{
\item{linguisticprs}{linguistic probabilities as a vector of characters}
}
\value{
vector of probabilities foreach state. -1 if error occurs.
}
\description{
Map node state belief probabilities, described as linguistic variables in rules, to 
probabilities. Normalises so probabilities sum to 1.
}
\examples{
none
}

