% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNSMILER_Public_Network.R
\name{newdbn}
\alias{newdbn}
\title{newdbn PUBLIC Constructor}
\usage{
newdbn(xdslfilepath)
}
\arguments{
\item{xdslfilepath}{- the file name plus path for the xdsl file to use}
}
\value{
dbn structure object
}
\description{
This function creates a DBNSMILER object from a SMILE xdsl file
}
\examples{
none
}

