% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNSMILER_Public_Get_Set.R
\name{num2chardate}
\alias{num2chardate}
\title{num2chardate}
\usage{
num2chardate(num, timeunit)
}
\arguments{
\item{num}{is the number of timeunits since 1900/1/1; can be a vector}

\item{timeunit}{- 'year','month','day','hour',minute','second'}
}
\value{
string date - %y, %y-%m, %y-%m-%d, %y-%m-%d-H, %y-%m-%d-%H-%M, %y-%m-%d-%H-%M-%S
format for 'year','month','day','hour',minute','second' timeunit respectively
}
\description{
Inverse of chardate2num - converts a numeric number of timeunits since 1900/1/1
to string. Timeunits are: years, months, days, hours, minutes, seconds
}
\examples{
none
}

