% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNSMILER_Public_Network.R
\name{renewdbn}
\alias{renewdbn}
\title{renewdbn}
\usage{
renewdbn(dbn)
}
\arguments{
\item{dbn}{object to renew (see def in DBNSMILER package)}
}
\value{
updated dbn object (see def in DBNSMILER package)
}
\description{
This function re-initialises the SMILE net object (clears all evidence etc.) by 
reloading from the xdsl file stored in dbn object.
ASSUMPTION: ASSUMES NO CHANGE TO XDSL FILE SINCE CALL TO newdbn
}
\examples{
none
}

