% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNSMILER_Public_Network.R
\name{restruct}
\alias{restruct}
\title{restruct}
\usage{
restruct(dbndata, dbnstruct)
}
\arguments{
\item{dbndata}{- a DBNSMILER object to merge (use the data in this object - $node$notes, 
thresh, rules; $datetype, $globdater, $evid, $pplot, $lineplots, $contourplots)}

\item{dbnstruct}{- a DBNSMILER object to merge (use the order of nodes in this object, and node 
structure - $node$name, $id, $states, $dtslices, $parnodenames, $parnodeinds, $parstates, $cpt)}
}
\value{
dbn - a dbn object
}
\description{
This function restructures a DBN data object according to another
}
\examples{
none
}

