% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNSMILER_Public_Network.R
\name{restructrules}
\alias{restructrules}
\title{restructrules}
\usage{
restructrules(rulesdata, rulesstruct)
}
\arguments{
\item{rulesdata}{- a rules table to merge (use the actual rules rows in this object)}

\item{rulesstruct}{- a rules table to merge (use the row headers and dimensions of this object)}
}
\value{
rules - a rules table, -1 if restructrules fails
}
\description{
This function restructures a rule table (data frame) according to another. Format of
rules table is as described in 'initrules'
}
\examples{
none
}

