% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNSMILER_Public_Batch.R
\name{runbatchdbn}
\alias{runbatchdbn}
\title{runbatchdbn}
\usage{
runbatchdbn(evid, dbn, evidpath, postpath, postpath2, pctiles, numclusters)
}
\arguments{
\item{evid}{is DBNSMILER evid data frame for scenarios: EvidID (int), $Scenario (char), 
$Node (name as str), $DateStart (char), $DateEnd (char), $SelectState (char), SoftEvidence (char)
see DBNSMILER-package documentation}

\item{dbn}{DBNSMILER object (see DBNSMILER-package documentation). Use dbn to run inference.}

\item{evidpath}{- path to local folder for storing evidence distributed data frame}

\item{postpath}{- path to local folder for storing posterior distributed data frame}

\item{postpath2}{- path to local folder for storing re-structured by scenario and node posterior
distributed data frame for visualisation}

\item{pctiles}{- percentiles to do weighted mean calculations with (e.g. c(0,0.25,0.5,0.75,1))}

\item{numclusters}{- number of clusters to use for processing (1=no parallel processing)}
}
\value{
List containing ddfevid, ddfp, ddfpost: posterior as distributed data frame (ddf-local 
disk). ddfp is divided as per evid ddf (drkeys), ddfpost is divided by scenario and node.
ddfevid is ddf version of evid divided by drkeys=scenario
Note the base dbn$posterior data frame struct is: $node,$state,$t,$p,$scenario,$pctile,$xbar
(see DBNSMILER-package documentation). Returns NA for both if errors.
}
\description{
Run a batch of DBN simulations according to specified evidence data frame.
Note this function uses datadr library.
}
\examples{
none
}

