% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNSMILER_Public_InfVis.R
\name{rundbn}
\alias{rundbn}
\title{rundbn}
\usage{
rundbn(dbn, evid, inftype, dtnode, datetype, rettype)
}
\arguments{
\item{dbn}{DBNSMILER object (see DBNSMILER-package documentation)}

\item{evid}{table (data frame) assumed to have th
e same struct as dbn$evid (see DBNSMILER-package documentation)  E$EvidID (int), 
$Scenario (char), $Node (name as str), $DateStart (char), $DateEnd (char), $SelectState (char),
SoftEvidence (char)}

\item{inftype}{- integer of the inference type, 3 is likelihood sampling 8 is EPIS algorithm}

\item{dtnode}{- char name (id) of datetime node (e.g. 'Time_of_Year'); set to null if unused}

\item{datetype}{'year','month','day','hour','minute','second'}

\item{rettype}{- 0 to return the dbn with dbn$posterior updated, 1 to just return posterior}
}
\value{
dbn updated dbn object after inference and updated with results/marginals
Note, only the SMILE DBN (via dbn$net) and dbn$posterior are updated
}
\description{
Runs the DBN.
Takes as input the selected set/subset of evidence E for a scenario 
NOTE - re-setting time slice should not be necessary, but bug with SMILE causes change 
to number of slices to 3 after set evid, regardless of what numslices was
}
\examples{
none
}

