% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNSMILER_Public_Get_Set.R
\name{runinf}
\alias{runinf}
\title{runinf}
\usage{
runinf(net)
}
\arguments{
\item{net}{- SMILE DBN object}
}
\value{
1 if success, -1 if fail
}
\description{
Run inference on DBN
}
\examples{
none
}

