% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNSMILER_Public_InfVis.R
\name{setevid}
\alias{setevid}
\title{setevid}
\usage{
setevid(E, net, globdater, datetype)
}
\arguments{
\item{net}{SMILE DBN object}

\item{globdater}{defines simulation window; $char 2 element of datetime, 
$num, 2-element of datenum}

\item{datetype}{'year','month','day','hour','minute','second'}

\item{evid}{table (data frame)}
}
\value{
1 for success, -1 for failure
}
\description{
Takes as input the selected set/subset of evidence E for a scenario; E is assumed to
have the same struct as evid (see DBNSMILER-package documentation)  E$EvidID (int), 
$Scenario (char), $Node (name as str), $DateStart (char), $DateEnd (char), $SelectState (char),
SoftEvidence (char).
NOTE: given the same node, evidence in later rows (i.e. lower in the table) will write over
evidence in earlier rows. For example, to set evid for X to be X1 at all times except at
t2-t3 set it to X2, do: [row 1 - set X=X1][row 2 - set X=X2 at t2-t3]
}
\examples{
none
}

