% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNSMILER_Public_Batch.R
\name{setevidbypattern}
\alias{setevidbypattern}
\title{setevidbypattern}
\usage{
setevidbypattern(E, scnid, nodename, start, end, pattern.t, pattern.y, timeunit)
}
\arguments{
\item{E}{is DBNSMILER evid data frame that is being added to: 
EvidID (int), $Scenario (char), $Node (name as str), 
$DateStart (char), $DateEnd (char), $SelectState (char), SoftEvidence (char); 
see DBNSMILER-package documentation}

\item{scnid}{is the scenario id}

\item{nodename}{= node name}

\item{pattern.t}{= time of year to apply evidence (numeric)}

\item{pattern.y}{= evidence to set (as char) - same length as pattern.t}

\item{timeunit}{= 'month' (currenly the only supported time unit)}

\item{start/end}{= start/end time (char - same format as DBNSMILER$globdater$char)}
}
\value{
evid table formatted/edited; returns input arg 'evid' if error check fails
}
\description{
Set evidence using an annual temporal pattern
}
\examples{
none
}

