% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNSMILER_Public_Get_Set.R
\name{setinftype}
\alias{setinftype}
\title{setinftype}
\usage{
setinftype(net, inftype)
}
\arguments{
\item{net}{- SMILE DBN object}

\item{inftype}{- integer of the inference type, 3 is likelihood sampling 8 is EPIS algorithm}
}
\value{
1 if success, -1 if fail
}
\description{
Set the type of inference to perform. NOTE SMILE does not throw error if inftype
is invalid
}
\examples{
none
}

