% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNSMILER_Public_Get_Set.R
\name{setnumtslices}
\alias{setnumtslices}
\title{setnumtslices}
\usage{
setnumtslices(net, numslices)
}
\arguments{
\item{net}{- SMILe net object}

\item{numslices}{- the number of time slices to set (this value needs to be big enough to cover
global date range)}
}
\value{
none
}
\description{
Set the number of time slices for the DBN
}
\examples{
none
}

