% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNSMILER_Public_Get_Set.R
\name{setnumtslicesfordbn}
\alias{setnumtslicesfordbn}
\title{setnumtslicesfordbn}
\usage{
setnumtslicesfordbn(dbn)
}
\arguments{
\item{dbn}{- DBNSMILER object (see definition in DBNSMILER-package)}
}
\value{
1 if success, -1 if fail
}
\description{
Set the number of time slices for the DBN from the DBN global date range
}
\examples{
none
}

