% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNSMILER_Public_Batch.R
\name{visbatchdbn}
\alias{visbatchdbn}
\title{visbatchdbn}
\usage{
visbatchdbn(ddfpost, vdbpath, plottype, dateformat, customiseplots)
}
\arguments{
\item{vdbpath}{- path to local folder for storing trelliscope visualisation}

\item{plottype}{- 'wmean', 'statebeliefs','both'}

\item{dateformat}{- char specifying x-axis date format, e.g. \%m-\%Y for months-year
\%d for day, \%H for hour, \%M for minute, \%S for second}

\item{customiseplots}{- if 1, looks for function customiseplotswrapper to customise plot colours}

\item{ddfpost:}{posterior as distributed data frame (local disk), divided by: scenario and node,
formatted as per dbn$posterior (see DBNSMILER-package documentation):
$node,$state,$t,$p,$scenario,$pctile,$xbar}

\item{dbn}{DBNSMILER object (see DBNSMILER-package documentation). Use dbn to run inference.}
}
\value{
vdbconn - the visualisation database connection
}
\description{
Visualise a batch of DBN simulations according to posterior from runbatchdbn.
Note this function uses datadr and trelliscope library.
}
\examples{
none
}

