% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNSMILER_Public_InfVis.R
\name{wmean}
\alias{wmean}
\title{wmean}
\usage{
wmean(P, states, statelow, stateupp, pctiles, Ithresh)
}
\arguments{
\item{P}{- subset of dbn$posterior (see DBNSMILER-package documentation) for a given:
node, scenario and time - therefore, nrow(P)=numstates of that node. P has
$node,$state,$t,$p,$scenario,$pctile,$xbar}

\item{states}{- node states as a vector of chars}

\item{statelow}{- vector of lower thresholds ASSUME same length as states and in same order}

\item{stateupp}{- vector of upper thresholds ASSUME same length as states and in same order}

\item{Ithresh}{- threshold for zero indicator; set as NA if not in use}

\item{pctiles-}{numeric vector of percentiles for evaluating expected value=weighted mean at}
}
\value{
posterior data frame populated with weighted mean results
}
\description{
Calculates the weighted mean=expected value for the general case - see calcxbar
ASSUME order of states is same as that stored in SMILE and same as the order of thresholds
As part of this assumption, ASSUME that the last state is the Zero state if it exists.
}
\examples{
none
}

